ANDROID: dma-buf: Add support for mapping buffers with DMA attributes
commit f57c5b53d9a29ac14a73d624b41c104c8141a444 ACK

When mapping the memory represented by a dma-buf into a device's
address space, it might be desireable to map the memory with
certain DMA attributes. Thus, introduce the dma_mapping_attrs
field in the dma_buf_attachment structure so that when
the memory is mapped with dma_buf_map_attachment, it is mapped
with the desired DMA attributes.

Bug: 133508579
Test: ion-unit-tests
Change-Id: Ib2e5bafdc02ae31a58ce96a82d77cc508dd71bd4
Signed-off-by: Isaac J. Manjarres <isaacm@codeaurora.org>
Signed-off-by: Sandeep Patil <sspatil@google.com>
Reviewed-on: https://gerrit-sha.synaptics.com/gerrit/258616
Reviewed-by: Jisheng Zhang <Jisheng.Zhang@synaptics.com>