ANDROID: dma-heap: Rework allocation calls to return struct dma_buf instead of fd
commit 349e8360c9c85c907147ea2d1d4716468409f2ab ACK

While most uses will want to allocate a fd for a dmabuf, there
are some cases where it might be useful to have just a dma_buf structure.

So refactor the heap allocaiton functions to return a struct dma_buf
and add a dma_heap_bufferfd_alloc() function to convert that return a fd.

NOTE: the cast bug fix[1] has been folded-in

[1] https://android-review.googlesource.com/c/kernel/common/+/3122512

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: Id6fd7e8471b9555a08bcdf8ca210feb589fa51c1
Bug: 154341375
Reviewed-on: https://gerrit-sha.synaptics.com/gerrit/258622
Reviewed-by: Jisheng Zhang <Jisheng.Zhang@synaptics.com>