drm/imx: dw_hdmi-imx: allow pixel clock tolerance to 10%
The HDMI PHY driver currently can only generate a discrete set of
pixelclock frequencies. Any mode read from a monitor's EDID with a non
matching frequency is rejected.

Change that to find the closest possible frequency and accept the mode
if it deviates less than 10%. 10% is chosen as with that any pixelclock
in the range of 22.25Mhz - 297MHz should be accepted.

Seen on a  i.MX 8MP SoCs. There are some resolutions e.g. 1280x1024@60
Hz that do not work without this commit.

The dw_hdmi-imx.c file is available upstream but they do not
verify the pixel clock frequency. Adding this check would harm more than it
solves.

Related-to: ELB-4728
Upstream-Status: Pending
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>