imx8mp: native hdmi: allow pixelclocks being slightly off
The HDMI PHY driver currently can only generate a discrete set of
pixelclock frequencies. Any mode read from a monitor's EDID with a non
matching frequency is rejected.

Change that to find the closest possible frequency and accept the mode
if it deviates less than 6%. 6% is chosen as with that any pixelclock
in the range of 22.25Mhz - 297MHz should be accepted.

Downstream specific, i.MX 8M Plus HDMI subsystem is on mainline kernel
but implementation is different.

Related-to: ELB-4015
Upstream-Status: Inappropriate [other]
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>