arm64: dts: imx8m{m,p}-verdin: set cma
Bring back CMA in device tree and set its size to 512MB for entire
Verdin family.

Overwrite the alloc-ranges property.
On modules with 4GB of RAM, the CMA configuration does not work well
if not limiting the alloc-ranges property. Some of the display
subsystem drivers are unable to access existing CMA memory ranges.

This solves the display not working issue on Verdin iMX8MM modules
with 4GB of RAM.

Upstream-Status: Inappropriate [downstream specific]
Signed-off-by: Andrejs Cainikovs <andrejs.cainikovs@toradex.com>
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>
Signed-off-by: Parth Pancholi <parth.pancholi@toradex.com>