NVIDIA: SAUCE: dmaengine: tegra: Get slave channel based on DT
BugLink: https://bugs.launchpad.net/bugs/2072591

Add option for DMA clients to provide channel number and
stream-id in device tree. dma-cells need to be increased to 3
for accommodating additional details.

This is required in a virtualized environment where all channels
and clients are not allowed use the same stream id.

The channel number read from device tree is modified to match with
the hardware channel number in the driver. This is done to avoid
ABI breakage. Otherwise, it would require updating all the device
trees that provide this information

http://nvbugs/200761027

Signed-off-by: Akhil R <akhilrajeev@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: Abhilash G <abhilashg@nvidia.com>
Tested-by: Abhilash G <abhilashg@nvidia.com>
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>