arm64: dts: apalis-imx8: fec: add fsl,mii-exclusive
When fsl,mii-exclusive is not specified the driver only uses the
MII bus of the fec device with dev_id=0 for both fec1 and fec2.

On the Apalis iMX8 hardware design a potential second PHY is
connected to the second MAC's MII bus.

This property is needed to select the MDIO bus coupled with fec2, if
fec2 is used. We put it in fec1 for customers so they automatically
inherit this property if they placed a PHY on their carrier-board
and copied the fec1 node.

Compare with commit 9801f2e913a2 ("net: fec: reinit MAC0 MII bus to
be used by MAC1 after resume back")

Mainline does not support the fsl,mii-exclusive property.

Upstream-Status: Inappropriate [Downstream only]
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>