gpu: drm: imx: lcdif-mux: enable the clk_pixel when the driver binds
Make sure to enable the clk_pixel clock during bind so that all other
driver in the display pipeline can rely on it. Without this change we
might see ghosting effects in around 10% of the reboots because
something in the display pipeline runs out of sync. With this change we
don't disable the clk_pixel when the display is turned off so this might
increase the power consumption in standby mode a bit. However, it would
require some bigger changes in the display pipeline to allow this.

The dpu is currently only supported in the downstream kernel. A fix will
most likely be applied by NXP to the downstream kernel. A similar patch
was provided by them but it breaks the DPMS off/standby mode. See this
ticket for more information:
https://support.nxp.com/s/case/5002p00002vUMOL/ghosting-effect-on-parallel-rgb-interface

Upstream-Status: Inappropriate [other]
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>