NVIDIA: SAUCE: ASoC: tegra: support few external codecs
BugLink: https://bugs.launchpad.net/bugs/2072591

Few external audio cards are supported out of the box. DT bindings
provide specific DAI link names for an audio card, which are then
parsed in driver to apply specific configuration.

However this solution is rejected in upstream, because the generic
machine driver is not supposed to include codec headers for macros
which are used in set_sysclk() and set_pll() callbacks.

Couple of attempts were previously made to fix this, but these
have been rejected as well:
  1. https://lore.kernel.org/all/1614276364-13655-1-git-send-email-spujar@nvidia.com/
  2. https://lore.kernel.org/lkml/1646912477-3160-1-git-send-email-spujar@nvidia.com/

In summary, the DT maintainers don't want to have another method to
provide clock relationship for codec sysclk. Suggestion is to re-use
existing standard clock bindings. The problem with standard bindings
is fixes will be needed in codec driver and it is not clear who is
going to own these changes. Unfortunately we need to maintain this
change in downstream until we have a standard solution.

DT binding reference:
	/*
	 * "link-name" is used to apply specific codec config.
	 * Driver parses this and applies necessary config at
	 * runtime.
	 */
	codec_port {
		endpoint {
			remote-endpoint = <&cpu_ep>;
			mclk-fs = <256>;
			link-name = "rt5640-codec-sysclk-bclk1";
		};
	};

http://nvbugs/3585083
http://nvbugs/4451662

Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Mohan kumar <mkumard@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>