gpu: imx: lcdifv3: power down imx8mp hdmi PHY
The lcdifv3 video controller has a clock dependency on the HDMI PHY for
the imx8mp. Powering down the HDMI PHY before the drm display pipeline
finishes after a hot-plug monitor disconnection event causes a vblank
timeout error due to the lcdif clock stopping. Likely because of this,
the downstream code doesn't power down the HDMI PHY (it was commented),
leaving the PHY ON even on a disconnection event, resulting in the HDMI
data and clock signals still running. This appears to cause issues when
using HDMI with bridge devices (e.g. HDMI-->eDP) to handle HPD (Hot-Plug Detect)
events from the bridge.

To address the problem, the workaround moves the PHY power-off logic to
the suspend of the lcdifv3 driver. At this position, it guarantees that
the DRM pipeline has finished and disabled and the PHY can be powered off
without problems. Doing this appears to be safe as the PHY is powered on
again on imx8mp_hdmi_phy_init() before starting the DRM pipeline. The PHY
is controlled by the HDMI TX BLK CTRL peripheral, but the downstream driver
for the peripheral is not fully implemented. So, the workaround maps
HDMI TX BLK CTRL register space to the lcdifv3 driver and controls it
directly, similar to the approach used in the imx HDMI driver.

The imx8mp already has a fully working implementation on upstream [1] and
[2]. Nonetheless, attempting to backport it appears impractical, leading
to the adoption of this workaround.

[1] https://lore.kernel.org/lkml/20240210204606.11944-1-aford173@gmail.com/
[2] https://lore.kernel.org/lkml/20240203165307.7806-1-aford173@gmail.com/

Upstream-Status: Inappropriate [other]
Signed-off-by: Joao Paulo Goncalves <joao.goncalves@toradex.com>