arm64: dts: freescale: imx8mm-verdin: remove ddrc node on downstream
Remove the ddrc node reference. Without removing it, the kernel doesn't compile
because there is no ddrc node to reference on the downstream DT of iMX8MM.

The NXP kernel is using the busfreq-imx8mq.c driver do control the ddr clocks.
This approach is different from upstream that uses the ddrc node with another
driver so this change is downstream only.

Upstream-Status: Inappropriate [other]
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>