NVIDIA: SAUCE: iommu/of: Fix pci_request_acs() before enumerating PCI devices
BugLink: https://bugs.launchpad.net/bugs/2120190

When booting with devicetree, the pci_request_acs() is called after the
enumeration and initialization of PCI devices, thus the ACS is not
enabled. And ACS should be enabled when IOMMU is detected for the
PCI host bridge, so add check for IOMMU before probe of PCI host and call
pci_request_acs() to make sure ACS will be enabled when enumerating PCI
devices.

Fixes: 6bf6c24720d33 ("iommu/of: Request ACS from the PCI core when configuring IOMMU linkage")
Signed-off-by: Xingang Wang <wangxingang5@huawei.com>
Signed-off-by: Pavankumar Kondeti <quic_pkondeti@quicinc.com>
[jniethe: applied from upstream mailing list:
https://lore.kernel.org/lkml/20241107-pci_acs_fix-v1-1-185a2462a571@quicinc.com/]

http://nvbugs/5217578

Signed-off-by: Jordan Niethe <jniethe@nvidia.com>
Reviewed-by: Pritesh Raithatha <praithatha@nvidia.com>
Signed-off-by: Brad Griffis <bgriffis@nvidia.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Acked-by: Abdur Rahman <abdur.rahman@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>