drm: imx: sec_mipi_dsim-imx: allow pixel clock rate deviation
In many cases the rounded_rate is not exactly the same as the
pixel_clock_rate, in other words, some small deviations occurs. This is
a snippet of some prints comparing both values when the edid is forced
to a fixed value in the kernel arguments:

[    9.525369] hfranco: pixel_clock_rate: 36000000
[    9.525393] hfranco: rounded_rate: 35844828
[    9.525470] hfranco: pixel_clock_rate: 40000000
[    9.525475] hfranco: rounded_rate: 39980770
[    9.525551] hfranco: pixel_clock_rate: 33750000
[    9.525555] hfranco: rounded_rate: 33532259
[    9.525628] hfranco: pixel_clock_rate: 65000000
[    9.525631] hfranco: rounded_rate: 64968750
[    9.732019] hfranco: pixel_clock_rate: 36000000
[    9.732039] hfranco: rounded_rate: 35844828
[    9.732118] hfranco: pixel_clock_rate: 40000000
[    9.732123] hfranco: rounded_rate: 39980770
[    9.732201] hfranco: pixel_clock_rate: 33750000
[    9.732210] hfranco: rounded_rate: 33532259
[    9.854640] hfranco: pixel_clock_rate: 36000000
[    9.854661] hfranco: rounded_rate: 35844828
[    9.854763] hfranco: pixel_clock_rate: 40000000
[    9.854768] hfranco: rounded_rate: 39980770
[    9.854869] hfranco: pixel_clock_rate: 33750000
[    9.854874] hfranco: rounded_rate: 33532259
[   12.847460] hfranco: pixel_clock_rate: 36000000
[   12.847479] hfranco: rounded_rate: 35844828

Therefore, allow a +/-0.5% pixel clock rate deviation to make the mipi
dsi converter properly works. This fixes an issue where weston is not
able to start because there is no mode available for HDMI.

This patch is downstream only, since this code and file does not exist
upstream.

Upstream-Status: Inappropriate [other]
Fixes: c8966cffa993 ("LF-7555-4: drm: imx: sec_mipi_dsim: Introduce pixel clock validation")
Related-to: TEI-1226
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>