arm64: dts: imx8-apalis: remove fsl,max-link-speed for pcie1
Removing fsl,max-link-speed property from imx8-apalis-v1.1.dtsi, which
limits the speed for the Apalis PCIe1 interface to Gen 1. The PCIe
interface of the Apalis iMX8 family can reach speeds up to Gen 3. There
is no reason to limit it artificially, as the link speed is negotiated.

This patch is downstream only, as the node for pcie1 does not exist
upstream.

Upstream-Status: Inappropriate [other]
Signed-off-by: Franz Schnyder <franz.schnyder@toradex.com>