arm64: dts: freescale: add initial support for Toradex SMARC iMX95
Add basic devicetree support for Toradex SMARC iMX95 SoM and SMARC development
carrier board.

Working basic features when using Verdin iMX95 EVK:
        * eMMC
        * debug UART

Not possible to test on the EVK:
        * Ethernet
	* USB

The devicetree for the development carrier board includes the module's DT
and enables the supported peripherals of the carrier board.

It will be upstreamed after bringup is completed.

Upstream-Status: Pending
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Co-developed-by: João Paulo Gonçalves <joao.goncalves@toradex.com>
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>