arm64: dts: freescale: imx95-toradex-smarc: Add SPI0, SPI1 and TPM
Add SMARC SPI0, SPI1 and SPI TPM. Pad control configuration (0x3fe) is
fast slew rate, x6 drive strength and pull-up enabled. Configuration
confirmed with our HW team and this is the same as used by NXP on their
evk DTS file.

On SMARC SPI1 chip selects are used as GPIO, this is done to allow
having the SMARC CS0# and CS1# mapped as cs0 (reg = <0>) and
cs1 (reg = <1>) in Linux, that would not be possible using the native CS
functionality. No drawbacks are expected from this decision.

The actual TPM used on the board is not well defined, as we have second
source options. Given that the driver does not need to know which part
is exactly assembled, to maximize flexibility, use the infineon
compatible, that is also the one requiring the slowest SPI clock. The
TPM node has a label, currently not used, to make it easier to override
such configuration if needed.

Upstream-Status: Pending
Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>