arm64: dts: freescale: imx8mp-toradex-smarc: update pinctrl comments
It has been decided to use the following pattern when adding the pinctrl
nodes: each pin gets a comment with the SMARC pin number (SMARC pin
name) if applicable, otherwise it gets a comment with the netname. We
will not add the ball number to the comment.

It will be upstreamed after bringup is completed.

Upstream-Status: Pending
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>