arm64: dts: freescale: imx8mp-toradex-smarc: disable pulls on I2S0 pins
The I2S is a push-pull bus, and as a result, it doesn't need a pull-up
resistor. This change disables the pull on I2S0 pads.

Also, it decreases the drive strength as the I2S is a low-speed protocol
and may introduce ringing on the signal.

It will be upstreamed once the bring-up process is completed.

Upstream-Status: Pending
Fixes: bd626d8c6e1c ("arm64: dts: freescale: imx8mp-toradex-smarc: add WM8904 audio codec")
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>