arm64: dts: freescale: imx8mp-toradex-smarc: fix ethernet mdio pinctrl
The MDIO driver does not configure the required pinctrl settings,
resulting in the PHYs not being recognized.

To resolve this issue, move the MDIO pinctrl configuration directly
under the EQOS node. With this change, both PHYs (EQOS and FEC) are
correctly initialized and functioning as expected.

It will be upstreamed once the bring-up process is completed.

Upstream-Status: Pending
Fixes: 6333e580b3e3 ("arm64: dts: freescale: add initial support for Toradex SMARC iMX8MP")
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>