arm64: dts: freescale: imx8mp-toradex-smarc: fix regulator-wifi-en node
When the SDHCI driver attempts to retrieve the VMMC and VQMMC regulators,
the `regulator-min-microvolt` and `regulator-max-microvolt` properties
must configured properly. Without these properties, the following warning
is raised:

  "sdhci-esdhc-imx 30b40000.mmc: could not set regulator OCR (-22)"

This commit resolves the issue by adding the missing properties to the
`regulator-wifi-en` node. After the fix, the usdhc on 30b40000 operates
as expected.

It will be upstreamed once the bring-up process is completed.

Upstream-Status: Pending
Fixes: 3c6d4097584d ("arm64: dts: freescale: imx8mp-toradex-smarc: add Wi-Fi/BT support")
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>