arm64: dts: freescale: imx8mp-toradex-smarc: Enable WP on SDIO SD card
Currently, the SDIO SD card interface does not have enabled the write
protection (WP) feature, even though the pinctrl is configured:

[  170.846393] mmc1: host does not support reading read-only switch, assuming write-enable

To address this, change the pin function to GPIO and add `wp-gpios`
property using it to the SDIO SD card interface (usdhc2).

Additionally, remove the `bus-width` property from the node configuration,
as it is already defined in the imx8mp.dtsi file, avoiding redundancy.

It will be upstreamed once the bring-up process is completed.

Upstream-Status: Pending
Fixes: 25d1b17922cf ("arm64: dts: freescale: imx8mp-toradex-smarc: add SDIO SD card support")
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>