arm64: dts: freescale: imx8mp-toradex-smarc: add vbus-supply to dwc3_phy1
The USB hub on the SoM was not working due to a missing `vbus-supply`
property in the `dwc3_phy1` node.

This change adds the `vbus-supply` property with the correct regulator
to resolve the issue. After this modification, the USB hub is functioning
as expected.

It will be upstreamed once the bring-up process is completed.

Upstream-Status: Pending
Fixes: 7ea5d163e1e4 ("arm64: dts: freescale: imx8mp-toradex-smarc: add USB devices")
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>