arm64: dts: freescale: imx8mp-toradex-smarc: fix PCIe reset property
The pci-imx6 driver parses the `reset-gpio` property from the DT for
the reset pin rather than requesting it from the GPIO subsystem. As
the current DT uses the `reset-gpios` property, the driver does
not recognize the configuration, and consequently, the PCIe endpoint
is not enumerated.

To address this, change the property name in the device tree to
`reset-gpio`, ensuring it is correctly parsed by the driver. After
this fix, the PCIe endpoint is successfully enumerated.

Upstream pci-imx6 driver uses `devm_gpiod_get_optional()` to request
the reset gpio, so there is no need to send this change.

Upstream-Status: Inappropriate
Fixes: 80321a585b38 ("arm64: dts: freescale: imx8mp-toradex-smarc: add PCIe support")
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>