arm64: dts: freescale: imx8mp-toradex-smarc: add LVDS-DSI MUX control
The SMARC_LVDS1 and SMARC_DSI1 signals are multiplexed, meaning they
cannot be used simultaneously. A GPIO pin is configured as a `gpio-hog`
to control which signal is active.

By default, the hardware is configured to use the DSI interface. In
software, when the `gpio-hog` is set to `output-high`, the LVDS
interface is enabled, and when set to `output-low`, the DSI interface is
enabled. This configuration change is best implemented using a device
tree overlay or by modifying the carrier board's device tree.

Upstream-Status: Pending
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>