arm64: dts: freescale: imx8mp-toradex-smarc: fix usdhc2 vqmmc
The current device tree is wrong since the IO voltage is controller
twice in parallel, once by regulator subsystem and once by VSELECT pin.

To fix this issue, configure the VSELECT pin as a GPIO and use it as a
GPIO regulator to properly supply the correct vqmmc voltage to the usdhc2
interface.

Additionally, rename the label reg_usdhc2_sd to reg_usdhc2_vmmc to enhance
code readability.

Upstream-Status: Pending
Fixes: 25d1b17922cf ("arm64: dts: freescale: imx8mp-toradex-smarc: add SDIO SD card support")
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>