hwmon: (amc6821) Add PWM polarity configuration with OF
Add support to configure the PWM-Out pin polarity based on the device
tree. The binding would allow also to configure the PWM period, this is
currently not implemented by the driver.

The driver has a module option to set the PWM polarity (normal=0,
inverted=1), when specified it always takes the precedence over the DT.

Upstream-Status: Submitted [https://lore.kernel.org/all/20250402102146.65406-2-francesco@dolcini.it/]
Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>