arm64: dts: freescale: imx8mp-toradex-smarc: Use SPI[12]_CS0# as GPIOs
Currently, the SPI[12]_CS0# signal toggles with the clock signal and shall
remain low for the entire SPI transaction. The cause of this behavior
is unclear, but other iMX8MP DT files configure the CS pin as a GPIO.

To align with this approach, change the ECSPI1_SS0 and ECSPI2_SS0 pin
functions to GPIO and adjust the `cs-gpios` property accordingly. With
this change, the CS signal stays low throughout the transmission.

Additionally, add the TPM_CS# gpio line name to the corresponding gpio.

Upstream-Status: Pending
Fixes: c5f02294d148 ("arm64: dts: freescale: imx8mp-toradex-smarc: add SPI0, SPI1 and TPM device")
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>