arm64: dts: freescale: imx95-toradex-smarc: configure dp lane reordering
The DSI to Display Port bridge connected its physical lanes in reverse
order to the SoM connectors pin.
Thus move the endpoint node to the SoM dtsi where the lanes reordering
is needed and add the property which makes the bridge reordering the lanes.

Keep the endpoint graph connection in the carrier board to allow for
carrier boards which do not make use of the DP bridge.

Upstream-Status: Pending
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>