drm/bridge: imx95-mipi-dsi: HACK: valid pixel clock unconditionally
With a DP connector as the last endpoint the limitation to only
support a set of pixel clock frequencies is not checked.
Due to that the graphic pipeline may configure a mode which is invalid
and an attached monitor stays black.

Test the pixel clock unconditionally of next bridge capabilities to
workaround this.

Upstream-Status: Pending
Fixes: ebab97775306 ("LF-10868 drm/bridge: imx95-mipi-dsi: Validate mode against magic clock rates if needed")
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>