NVIDIA: SAUCE: cpufreq: tegra194: Support read-only initialization
BugLink: https://bugs.launchpad.net/bugs/2136282

If a platform uses system registers for frequency readback, support
initialization without a mapped MMIO space. This will prevent frequency
changes by the driver, but still allow for frequency readback.
This is important & useful in the automotive space where we want to
support platforms with frequency information, but disallow changes to
the frequency for safety purposes.

http://nvbugs/5213905
http://nvbugs/4824989

Signed-off-by: Sumit Gupta <sumitg@nvidia.com>
Signed-off-by: Ishan Shah <ishah@nvidia.com>
Reviewed-by: Nathan Hartman <nhartman@nvidia.com>
Reviewed-by: Brad Griffis <bgriffis@nvidia.com>
Signed-off-by: Prathamesh Shete <pshete@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>