drm/bridge: imx95-mipi-dsi: Handle concurrent DSI / LDB use
Move DSI to the second display output if the first display output
is configured to be used by the LDB (LVDS) bridge.
If both DSI and LDB display output is configured from the device
tree without this patch LDB wins and the DSI output is not used
at all.

The patch is provided by NXP, the driver is downstream only.

Upstream-Status: Inappropriate [other]
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>