arm64: dts: freescale: imx95-verdin: add usb recovery gpio hog
The reset configuration allows using Verdin USB_1 as the recovery port,
while the boot rom is executing.
After the recovery phase ends it is needed to change the USB
configuration to the default position.
Add USB_RECOV_CTRL# GPIO hog to switch USB routing to the default
position.

Upstream-Status: Pending
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>