NVIDIA: SAUCE: pstore: Avoid using pr_cont
BugLink: https://bugs.launchpad.net/bugs/2136282

This is done as pr_cont is generally discouraged in kernel code
because it can cause message interleaving in SMP systems.

This is also causing the print message to be split across lines
making it meaningless to interpret.

Solve this by using a buffer of size 256 bytes to print the text
and use pr_info to finally print all text together.

This buffer size is more than sufficient to avoid any buffer
overflow.

http://nvbugs/5277482

Signed-off-by: Bharat Nihalani <bnihalani@nvidia.com>
Signed-off-by: Kartik Rajput <kkartik@nvidia.com>
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Signed-off-by: Prathamesh Shete <pshete@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Noah Wager <noah.wager@canonical.com>
Signed-off-by: Noah Wager <noah.wager@canonical.com>