arm64: dts: freescale: imx95-toradex-smarc: Remove passive trip points
Remove the passive trip points from the thermal_zones node because they
are not mapped to any cooling action.

Passive (and active) trip points are supposed to be mapped to a cooling
action via the "cooling-map" node. Without this mapping, a passive trip
point does nothing.

Keep the critical trip points because they will shut down the system
when the PMICs reach the set critical temperature.

Upstream-Status: Pending
Fixes: 3fa909975495 ("arm64: dts: freescale: imx95-toradex-smarc: Add thermal zones")
Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>