drm/imx: imx8mp-ldb: Remove fixed pixel clock constraint
Remove the limitation of the ldb driver for users to be able to use
different pixel clocks. With this, it is now possible for users to
set up their displays without having to build a custom kernel.

The ldb driver forces all pixel clock settings to 74.25MHz,
or 148.5MHz for dual channel, with the justification that the PLL is
limited. Our testing shows that the PLL can generate more clocks despite
what the code comments seem to indicate.

We still need to consider that the code is limited: the reconfiguration of
the clock tree is not done automatically based on the pixel clock defined
on the panel node. It's necessary to manually change the
assigned-clock-rate of the IMX8MP_VIDEO_PLL1 clock following the formula:

clk-rate = pixel-clock * 7

The upstream driver is implemented differently and doesn't include the
code for the clock limitation.

Upstream-Status: Inappropriate [Other]
Signed-off-by: Rafael Beims <rafael.beims@toradex.com>