remoteproc: imx_rproc: skip clock enable when M-core is managed by the SCU
For the i.MX8X and i.MX8 family SoCs, when the M-core is powered up
by the bootloader, M-core and Linux are in same SCFW (System Controller
Firmware) partition, so linux has permission to control M-core.

But when M-core is started, the SCFW will automatically enable the clock
and configure the rate, and any users that want to enable the clock will
get error 'LOCKED' from SCFW. So current imx_rproc.c probe function
fails because clk_prepare_enable also fails. With that, the M-core power
domain is powered off when it is still running, causing a SCU (System
Controller Unit) fault reset, and the system restarts.

To address the issue, ignore handling the clk for i.MX8X and i.MX8 M-core,
because SCFW will automatically enable and configure the clock.

The submitted patches reached v5, using the generic power domain to get
the current power status of the remote core. However this v5 series
depends on a large patch series [1] that cannot be backported to the 6.6
dowsntream NXP kernel. Instead, let's use the v3 patch series [2], since
this approach also works. It reads the power status using the SCU API
and it can be backported to this kernel.

[1] https://lore.kernel.org/all/20250523134025.75130-1-ulf.hansson@linaro.org/
[2] https://lore.kernel.org/lkml/20250519171514.61974-1-hiagofranco@gmail.com/

Upstream-Status: Submitted [https://lore.kernel.org/lkml/20250617193450.183889-1-hiagofranco@gmail.com/]
Suggested-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>