remoteproc: imx_rproc: add power mode check for remote core attachment
When the remote core is started before Linux boots (e.g., by the
bootloader), the driver currently is not able to attach because it only
checks for cores running in different partitions. If the core was kicked
by the bootloader, it is in the same partition as Linux and it is
already up and running.

This adds power mode verification through the SCU interface, enabling
the driver to detect when the remote core is already running and
properly attach to it.

The submitted patches reached v5, using the generic power domain to get
the current power status of the remote core. However this v5 series
depends on a large patch series [1] that cannot be backported to the 6.6
dowsntream NXP kernel. Instead, let's use the v3 patch series [2], since
this approach also works. It reads the power status using the SCU API
and it can be backported to this kernel.

[1] https://lore.kernel.org/all/20250523134025.75130-1-ulf.hansson@linaro.org/
[2] https://lore.kernel.org/lkml/20250519171514.61974-1-hiagofranco@gmail.com/

Upstream-Status: Submitted [https://lore.kernel.org/lkml/20250617193450.183889-1-hiagofranco@gmail.com/]
Suggested-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>