verdin-imx95: add I2S_2_D_IN and SYNC as GPIO muxing
Add muxing for these two pins as GPIO. They are used as GPIOs from the
native LVDS overlay, specifically for the touch controller.

The I2S_2_D_IN GPIO muxing specifies a pullup instead of a pulldown
because this line is used for the active low reset of the touch
controller.

Will be upstreamed when device bringup is complete.

Upstream-Status: Pending
Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>