arm64: dts: freescale: imx95-toradex-smarc: add fan cooling levels
The fan controller on this board cannot work in automatic mode, and
requires software control, the reason is that it has no temperature
sensor connected.

Given that this board is a development kit and does not have any
specific fan, add a default single cooling level that would enable the
fan to spin with a 100% duty cycle, enabling a safe default.

Upstream-Status: Pending
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>