arm64: dts: freescale: imx8mp-toradex-smarc: fix lvds dsi mux gpio
The MUX which either outputs DSI or 2nd channel LVDS signals is part of
the SoM. Move the pinmuxing of the GPIO used for controlling the MUX
to the SoM dtsi file.

Change <tab> to a <space> on `gpio-line-names = "ETH_0_INT#", /* 0 */`
to resolve a merge conflict. This makes downstream use the same as
mainline.

Upstream-Status: Submitted [https://lore.kernel.org/all/20250708135300.2855155-1-max.oss.09@gmail.com/]
Fixes: 97dc91c04558 ("arm64: dts: freescale: add Toradex SMARC iMX8MP")
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>