Revert "remoteproc: imx_rproc: merge ITCM and DTCM regions"
This reverts commit b409ac306fb6a77f37d8cfdbea67126d1c5ae1bf.

Altough these are contiguous memories from the Cortex-A side, they are
not contiguous from the Cortex-M side:

- Cortex-A memory map:
	- ITCM: 0x7e0000 - 0x7fffff
	- DTCM: 0x800000 - 0x81ffff
- Cortex-M memory map:
	- ITCM: 0x0 - 0x1ffff
	- DTCM: 0x20000000 - 0x2001ffff

Therefore any older firwmare that is expecting the data section to be
inside the DTCM on 0x20000000 will not work anymore, since the
translation table removed this address.

This is a broken patch sent to the Linux kernel mailing list [1] which
will not be merged.

[1] https://lore.kernel.org/all/20250703130831.18830-1-hiagofranco@gmail.com/

Upstream-Status: Inappropriate [Other]
Related-to: ELB-6498
Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>