firmware: se_fw: fix memory reservation
The security enclave firmware runs on the Cortex-M33 core, and the driver
expects responses to requests to be returned via a shared memory region
in DRAM.

However, the Cortex-M33 cannot access the entire DDR address space.
If the driver uses an inaccessible region, the Cortex-M33 may fail
to respond to Cortex-A requests.

Fix this by enabling memory reservation in the driver and supporting the
`memory-region` property in the device tree to specify a suitable region.

Adapted to the lf-6.6.52-2.2.1 kernel update.
(cherry picked from commit af14377659e3fb5f107d7082703ecd74a4708388)

Upstream-Status: Pending
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>